Preserve 64-bit integer precision in value reads
Swift.Int/UInt/Int64/UInt64 reads truncated through NativePointer.readS64().toNumber(),
silently losing precision above 2^53. Return the Int64/UInt64 wrappers instead so reads
are exact; writers accept number|Int64/UInt64 and swiftValueEquals compares the wrappers
(two unequal Int64s previously compared equal via the empty-Object.keys path).

+++ tests/extended-existential.test.ts
@@ -29,7 +29,7 @@ describe("readValue extended existential", () => {
     const Holder = existentialMetadata("fixture.holderIntType");
     expect(Holder.kind).toBe(MetadataKind.ExtendedExistential);
     expect(extendedExistentialSpecialKind(Holder)).toBe(ExtendedExistentialSpecialKind.None);
-    expect(readValue(Holder, store("fixture.storeHolderInt", Holder))).toEqual({ item: 42 });
+    expect(readValue(Holder, store("fixture.storeHolderInt", Holder))).toEqual({ item: int64(42) });
   });
 
   test("decodes a class reference behind a class-constrained parameterized existential", () => {
@@ -39,7 +39,7 @@ describe("readValue extended existential", () => {
     expect(extendedExistentialSpecialKind(Ref)).toBe(ExtendedExistentialSpecialKind.Class);
     const ref = readValue(Ref, store("fixture.storeRefInt", Ref)) as NativePointer;
     expect(ref.isNull()).toBe(false);
-    expect(readObject(ref)).toEqual({ value: 7 });
+    expect(readObject(ref)).toEqual({ value: int64(7) });
   });
 
   test("decodes the stored type behind a parameterized-protocol existential metatype (any P.Type)", () => {
